requirements.txt
selenium
telethon
beautifulsoup4
pyyaml
chromedriver-autoinstaller